Use EKS and the EKS worker AMIs, rather than self-managing
**This commit does not leave the repository in a working state:
ingress is broken**

I've been having networking issues with the NixOS managed Kubernetes
and, after spending a week and a half on it, have given up.  It seems
like a race condition in cluster set-up, because sometimes when I
provision a cluster it'll just be working (and stay working), and
sometimes it'll just be broken (and stay broken).

---

So, what are the alternatives?

The main one seems to be [kops][], but it sounds like it relies a lot
on using the kops tool to configure stuff, and then for the user to
export the corresponding yaml (or terraform).  I don't much like that,
I like having config files which are the source of truth, which aren't
just some dump of the current state.  That'd be like defining all your
infrastructure by writing tfstate JSON files: a huge pain.

Another popular alternative is [EKS][], the AWS managed Kubernetes
service.  However, all it manages are the master nodes: you're freed
from needing to configure that side of things, but the workers are
still your responsibility.  But all is not lost for EKS!  AWS have an
AMI for EKS-optimised Kubernetes worker nodes, which is perfect.  With
master nodes managed by EKS and worker nodes configured by the
official AMI, you effectively have a managed Kubernetes cluster: true,
you need to scale workers yourself, and Amazon won't help you if your
workers run out of resources or whatever, but Amazon *do* have a
vested interest in making the AMI work well and need little to no
manual intervention.

EKS more fits how I like to work than kops, and I desperately want to
stop self-managing Kubernetes networking... so I've decided to go with
EKS + the worker AMI.

[kops]: https://github.com/kubernetes/kops
[EKS]: https://aws.amazon.com/eks/