Use ALBs for ingress
With the workers being inside an ASG, it connecting to NodePorts
doesn't really work, as I'd have to manually open a bunch of ports in
an ELB.  Thankfully, there is an ingress controller which
automatically provisions ALBs and sets up DNS so that everything just
works:

https://github.com/kubernetes-sigs/aws-alb-ingress-controller/

Gone is the big list of ports which has to be kept in sync in the
service files and the nginx config.  Services are assigned an
arbitrary NodePort, and the ingress controller tells the ALB where to
look.

---

One consequence of using ALBs is that every app needs a healthcheck
endpoint.  Currently these endpoints are defined in the ingress
resources, but they could perhaps be split out to a separate yaml file
and used to power some monitoring as well.

---

This also switches from nginx to caddy because nginx has some
interesting ideas about hostnames:

1. If given statically, nginx resolves all hostnames at start-up.

  - If a hostname doesn't resolve, nginx doesn't start.  This is a
    problem because the ALBs won't exist until the apps are deployed.

  - If a hostname is changed to point to a different IP, nginx won't
    pick up that change either.

2. If given dynamically (in a variable), nginx forces you to construct
   the proxy URL yourself.

  - It seems exceedingly difficult to construct a URL based on
    user-supplied input which doesn't open a possibility of SSRF.

Caddy solves both of these problems by just behaving sensibly with
hostnames.